% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bottleneck_distance}
\alias{bottleneck_distance}
\title{Bottleneck distance between two persistence diagrams}
\usage{
bottleneck_distance(x, y, delta = 0)
}
\arguments{
\item{x}{A matrix of shape \eqn{n \times 2} specifying the first persistence
diagram.}

\item{y}{A matrix of shape \eqn{m \times 2} specifying the second
persistence diagram.}

\item{delta}{A numeric value specifying the relative error. Defaults to
\code{0.0} in which case the exact Bottleneck distance is computed. If \code{delta}
is greater than \code{0.0}, an approximate Bottleneck distance is computed.}
}
\value{
A numeric value storing the Bottleneck distance between the two
persistence diagrams.
}
\description{
This function computes the Bottleneck distance between two persistence
diagrams of the same homology dimension. The diagrams must be represented as
2-column matrices. The first column of the matrix contains the birth times
and the second column contains the death times of the points.
}
\examples{
bottleneck_distance(
  persistence_sample[[1]]$pairs[[1]],
  persistence_sample[[2]]$pairs[[1]]
)

}
