% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence-set-class.R
\name{persistence-set}
\alias{persistence-set}
\alias{as_persistence_set}
\alias{format.persistence_set}
\alias{print.persistence_set}
\title{An 'S3' class object for storing sets of persistence diagrams}
\usage{
as_persistence_set(x)

\method{format}{persistence_set}(x, ...)

\method{print}{persistence_set}(x, ...)
}
\arguments{
\item{x}{A list of objects of class \link{persistence}.}

\item{...}{Additional arguments passed to the function.}
}
\value{
An object of class 'persistence_set' containing the set of
persistence diagrams.
}
\description{
An 'S3' class object for storing sets of persistence diagrams
}
\examples{
# Create a persistence set from a list of persistence diagrams
as_persistence_set(persistence_sample[1:10])
}
