% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence-class.R
\name{persistence}
\alias{persistence}
\alias{as_persistence}
\title{An \code{S3} class for storing persistence data}
\usage{
as_persistence(x, ...)
}
\arguments{
\item{x}{Persistence data to be transformed; either a \eqn{\geq 3}-column
matrix (or object coercible to one) with dimension/degree, start/birth, and
end/death columns, a list whose first element is such an object, an object
of class 'PHom' as returned by \verb{[ripserr::vietoris_rips()]}, or (a list as
returned by a \verb{*Diag()} function in \strong{TDA} (e.g. \verb{[TDA::ripsDiag()]})
whose first element is) an object of class 'diagram'.}

\item{...}{Parameters passed to methods.}

\item{degree}{Non-negative integer; the homology degree for which to recover
a matrix of persistence pairs.}

\item{modulus, max_dim, threshold}{Possibly missing parameters of the
calculation of the persistence data \code{x} to be included in the 'persistence'
object.}
}
\value{
Persistence data in the form of a list of 2-column matrices
containing birth-death pairs.
}
\description{
The \code{\link{persistence}} class is a list of 2-column matrices containing
birth-death pairs in each dimension/degree, starting with zero.
}
