% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phutil-package.R
\docType{package}
\name{phutil-package}
\alias{phutil}
\alias{phutil-package}
\title{phutil: Persistence Homology Utilities}
\description{
A low-level package for hosting persistence data. It is part of the 'TDAverse' suite of packages, which is designed to provide a collection of packages for enabling machine learning and data science tasks using persistent homology. Implements a class for hosting persistence data, a number of coercers from and to already existing and used data structures from other packages and functions to compute distances between persistence diagrams. A formal definition and study of bottleneck and Wasserstein distances can be found in Bubenik, Scott and Stanley (2023) \doi{10.1007/s41468-022-00103-8}. Their implementation in 'phutil' relies on the 'C++' Hera library developed by Kerber, Morozov and Nigmetov (2017) \doi{10.1145/3064175}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tdaverse/phutil}
  \item \url{https://tdaverse.github.io/phutil/}
  \item Report bugs at \url{https://github.com/tdaverse/phutil/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

Authors:
\itemize{
  \item Aymeric Stamm \email{aymeric.stamm@cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})
  \item Jason Cory Brunson \email{cornelioid@gmail.com} (\href{https://orcid.org/0000-0003-3126-9494}{ORCID})
}

Other contributors:
\itemize{
  \item Michael Kerber (HERA C++ code) [contributor]
  \item Dmitriy Morozov (HERA C++ code) [contributor]
  \item Arnur Nigmetov (HERA C++ code) [contributor]
}

}
\keyword{internal}
