% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wasserstein_distance}
\alias{wasserstein_distance}
\title{Wasserstein distance between two persistence diagrams}
\usage{
wasserstein_distance(x, y, wasserstein_power = 1, delta = 0.01)
}
\arguments{
\item{x}{A matrix of shape \eqn{n \times 2} specifying the first persistence
diagram.}

\item{y}{A matrix of shape \eqn{m \times 2} specifying the second
persistence diagram.}

\item{wasserstein_power}{A numeric value specifying the power of the
Wasserstein distance. Defaults to \code{1.0}.}

\item{delta}{A numeric value specifying the relative error. Defaults to
\code{0.01}.}
}
\value{
A numeric value storing the Wasserstein distance between the two
persistence diagrams.
}
\description{
This function computes the Wasserstein distance between two persistence
diagrams of the same homology dimension. The diagrams must be represented as
2-column matrices. The first column of the matrix contains the birth times
and the second column contains the death times of the points.
}
\examples{
wasserstein_distance(
  persistence_sample[[1]]$pairs[[1]],
  persistence_sample[[2]]$pairs[[1]]
)

}
