% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arch_spirals}
\alias{arch_spirals}
\title{A sample of persistence diagrams from the arch spiral}
\format{
A list of length 24, where each element is an object of class 'persistence'.
}
\usage{
arch_spirals
}
\description{
A collection of 24 samples of size 120 on the arch spiral from which a
persistence diagram is computed using the \code{\link[TDA:ripsDiag]{TDA::ripsDiag()}} function with
\code{maxdimension = 2} and \code{maxscale = 6}. Each diagram has been generated using
the \code{\link[tdaunif:arch-spirals]{tdaunif::sample_arch_spiral()}} function with the following parameters:
\code{n = 120}, \verb{arms = 2`` and }sd = 0.05`. The seed was fixed to 28415.
}
\keyword{datasets}
