% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{distances}
\alias{distances}
\alias{bottleneck_distance}
\alias{wasserstein_distance}
\title{Distances between two persistence diagrams}
\usage{
bottleneck_distance(x, y, delta = 0.01)

wasserstein_distance(x, y, delta = 0.01, wasserstein_power = 1)
}
\arguments{
\item{x}{A matrix of shape \eqn{n \times 2} specifying the first persistence
diagram.}

\item{y}{A matrix of shape \eqn{m \times 2} specifying the second
persistence diagram.}

\item{delta}{A numeric value specifying the relative error. Defaults to
\code{0.01}. For the Bottleneck distance, it can be set to \code{0.0} in which case
the exact Bottleneck distance is computed, while an approximate Bottleneck
distance is computed if \code{delta > 0.0}.}

\item{wasserstein_power}{A numeric value specifying the power of the
Wasserstein distance. Defaults to \code{1.0}.}
}
\value{
A numeric value storing the Bottleneck distance between the two
persistence diagrams.
}
\description{
This collection of functions computes the distance between two persistence
diagrams of the same homology dimension. The diagrams must be represented as
2-column matrices. The first column of the matrix contains the birth times
and the second column contains the death times of the points.
}
\examples{
bottleneck_distance(
  persistence_sample[[1]]$pairs[[1]],
  persistence_sample[[2]]$pairs[[1]]
)

wasserstein_distance(
  persistence_sample[[1]]$pairs[[1]],
  persistence_sample[[2]]$pairs[[1]]
)

}
