% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{pairwise-distances}
\alias{pairwise-distances}
\alias{bottleneck_pairwise_distances}
\alias{wasserstein_pairwise_distances}
\title{Pairwise distances within a set of persistence diagrams}
\usage{
bottleneck_pairwise_distances(x, delta = 0.01, ncores = 1L)

wasserstein_pairwise_distances(
  x,
  delta = 0.01,
  wasserstein_power = 1,
  ncores = 1L
)
}
\arguments{
\item{x}{A list of 2-column matrices specifying the set of persistence
diagrams.}

\item{delta}{A numeric value specifying the relative error. Defaults to
\code{0.01}. For the Bottleneck distance, it can be set to \code{0.0} in which case
the exact Bottleneck distance is computed, while an approximate Bottleneck
distance is computed if \code{delta > 0.0}.}

\item{ncores}{An integer value specifying the number of cores to use for
parallel computation. Defaults to \code{1L}.}

\item{wasserstein_power}{A numeric value specifying the power of the
Wasserstein distance. Defaults to \code{1.0}.}
}
\value{
An object of class 'dist' containing the pairwise distance matrix
between the persistence diagrams.
}
\description{
This collection of functions computes the pairwise distance matrix between
all pairs in a set of persistence diagrams of the same homology dimension.
The diagrams must be represented as 2-column matrices. The first column of
the matrix contains the birth times and the second column contains the death
times of the points.
}
\examples{
# Extract the list of 2-column matrices for dimension 0 in the sample
x <- lapply(persistence_sample[1:10], function(x) x$pairs[[1]])

# Compute the pairwise Bottleneck distances
Db <- bottleneck_pairwise_distances(x, delta = 0.01)

# Compute the pairwise Wasserstein distances
Dw <- wasserstein_pairwise_distances(x, delta = 0.01, wasserstein_power = 1.0)

}
