% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{persistence_sample}
\alias{persistence_sample}
\title{Toy Data: A sample of persistence diagrams}
\format{
A list of length 100, where each element is an object of class 'persistence'.
}
\usage{
persistence_sample
}
\description{
A collection of 100 samples of size 100 on the sphere from which a
persistence diagram is computed using the \code{\link[TDA:ripsDiag]{TDA::ripsDiag()}} function with
parameters \code{maxdimension = 1L} and \code{maxscale = 1.6322}. Each diagram has been
generated using the \code{\link[tdaunif:sphere]{tdaunif::sample_2sphere()}} function with the following
parameters: \code{n = 100} and \code{sd = 0.05}. The seed was fixed to 1234.
}
\keyword{datasets}
