% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phutil-package.R
\docType{package}
\name{phutil-package}
\alias{phutil}
\alias{phutil-package}
\title{phutil: Persistence Homology Utilities}
\description{
Implements a class for hosting persistence homology information in the form of a persistence diagram. The package also provides a set of functions to manipulate persistence diagrams, such as coercing them to different formats for compatibility with other packages or plotting them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tdaverse/phutil}
  \item \url{https://tdaverse.github.io/phutil/}
  \item Report bugs at \url{https://github.com/tdaverse/phutil/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

}
\keyword{internal}
