% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trefoils}
\alias{trefoils}
\title{A sample of persistence diagrams from the trefoil}
\format{
A list of length 24, where each element is an object of class 'persistence'.
}
\usage{
trefoils
}
\description{
A collection of 24 samples of size 120 on the trefoil from which a
persistence diagram is computed using the \code{\link[TDA:ripsDiag]{TDA::ripsDiag()}} function with
\code{maxdimension = 2} and \code{maxscale = 6}. Each diagram has been generated using the
\code{\link[tdaunif:trefoil]{tdaunif::sample_trefoil()}} function with the following parameters: \code{n = 120} and \code{sd = 0.05}. The seed was fixed to 28415.
}
\keyword{datasets}
